Code refactored and main function removed
The code is now ready to be merged into the master branch.
